fix(model-providers): own-property scope lookup; split test file under 300 lines (#5837)
Two CodeRabbit findings, both verified correct by executing the code.

- `rankOf` used `scopeType in SCOPE_RANK`, and `in` walks the prototype
  chain — so `"toString" in SCOPE_RANK` is true, `SCOPE_RANK["toString"]`
  returns a function, `Math.min(fn)` is NaN, and the NaN-comparator hole the
  `in` guard was added to close re-opens through inherited keys
  (toString/valueOf/constructor/__proto__). Measured: a row scoped
  "toString" WON the key over a PROJECT row, in both orders. Now
  `Object.prototype.hasOwnProperty.call(SCOPE_RANK, scopeType)`. Regression
  test asserts the semantic property (PROJECT wins), not order-independence
  — order-independence is TRUE while the bug is live (2 rows can't form an
  intransitive cycle), so an order test would be vacuous. Sabotage-verified.

- The resolution test file was a NEW file at 560 source lines, ~2x the
  repo's 300-line ceiling (.coderabbit.yaml:329, SRP). Split by resolver
  concern into precedence (229) / scopeRank (171) / malformedEntries (96) /
  labelKeys (95), all under the limit; makeProvider extracted to a shared
  test-helpers.ts, which also removes the duplication a hygiene review
  flagged against customModelDisplayNames.unit.test.ts. Pure move — no test
  behaviour changed; 9/9 scenario bindings preserved.

Refs #5837
